pyproject.toml
[tool.poetry]
name = "test_kafka"
version = "0.1.0"
description = ""
authors = ["qzibet <qzibet@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
kafka-python = "^2.0.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
